contrastive-fruits/requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy
Pillow
tqdm
opencv-python
albumentations
timm
matplotlib
scikit-learn
flask
requests
ultralytics